+++ ChangeLog.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- Add Annotations API as a service
+
 ## [v0.3.3] - 2023-06-28
 ### Security
 - Upgraded signature to match the security standard.